Got BrainPortal app and jiv app to communicate.
BrainPortal changes:
-added a list action to manager that sends a list of all available files in
xml.
-changed the content action in manager to only send the file, not the metadata

jiv changes:
-now uses Brainportal app to get list of images for a particular user(an image 
being any <name> such that <name>.header and <name>.raw_byte exist on the file
server.
-uploads files that are not chached from BrainPortal.

TODO:
-the way the apps exchange user information is very hacky. Right now
BrainPortal sends user_id in the uri. This has to be fixed.
-jiv currently does not authenticate. Need to add an action to BrainPortal to 
check that the current user is authenticated, so that jiv can use it.
-change call to manager/list into a post, so it can get user information 
(currently, it just sends all files that are available and jiv filters them).




git-svn-id: svn+ssh://huia.bic.mni.mcgill.ca/home/cbrain/cbrain_repo/trunk@20 bcc1f7e1-216c-4ce3-888e-c5a8bb225047

Former-commit-id: 4f75c9bfd3e8b1db890314d3940520de5ce96dca