New datalad data provider code.
The caching system and overall architecture was
redesigned to properly handle any subpath that
is registered as a userfile. (The previous code
only worked when under a dl repo was another dl repo)

A new feature is that the admin can elect to cache
the DP repo using the CBRAIN caching system local
on each RemoteResource, or to register a DataladSystemSubset
as a standard userfile to hold the cache (and in this
case it will benefit from the CBRAIN caching system
for transfering a copy of the repo between RRs)

This should be backwards compatible with existing
DataladDataProviders configured.