Patched the command for bigbrainwarp
System error, maybe related to Apptainer. Probably
not needed in the future and could be reverted.

If a bash script is located on the SquashFS filesystem
of a container image AND does not end properly with
a newline (like in this tool), the linux loader
will generate a Input/Output error and the script will
return a status code of 2.